Tuple toString (#10645)
* Use formattedWrite instead of format

* Remove workaround for shared objects in formattedWrite

The tests are still working. Is it safe to assume that the workaround
would've only been merged if there is a test-case checking it?

* Remove sharedToString workaround

For shared classes, it prints the mangled name, for struct pointers it will
print the adress.

This version of Tuple.toString won't work anymore in @safe code in
general if the tuple contains anything descended from Object. I think
this is correct even if a breaking change, since dealing with this issue
should be the responsibility of the caller.

* Comment to existing code

* Reference the bug that a unit test exemplifies