Add TrueTypeOf and ToFunctionType to phobos.sys.traits.
TrueTypeOf is completely new. It's a wrapper around typeof which
bypasses typeof's behavior with @property and instead gives the actual
type of the function. For other symbols, it's the same as typeof with
the caveat that template alias parameters cannot accept general
expressions. So, TrueTypeOf will work on symbols, but not stuff like
&var. But it will work on the symbol for a function or variable, thereby
providing a replacement for typeof in situations where property
functions need to be treated as the functions that they are.

The plan here is that the function-related traits in phobos.sys.traits
will operate solely on types (except in situations where the actual
symbol is required - e.g. to get the names of parameters), like most
traits typically do. std.traits has a number of its function-related
traits operate on symbols in part to work around the issue with property
functions (as well as stuff like trying to treat variables with opCall
as functions instead of requiring that the symbol for opCall itself be
passed). TrueTypeOf will therefore allow code to get the actual type of
the function and thus we won't need to have as many traits operate on
symbols instead of types. That will also have the benefit of reducing
the number of cases where traits operate on both types and other
symbols, since that has a tendency to be error-prone and make the code
harder to understand. It should also help with clarity when a trait
doesn't try to handle everything itself.

Exactly how this will affect each trait will of course depend on the
trait in question, but the idea is to simplify things and ultimately end
up with traits which are easier to understand and less error-prone.

The other new symbol, ToFunctionType, is a template which converts
function types, function pointer types, and delegate types to the
corresponding function type. So, something like `int function(string)`
or `int delegate(string)` would become `int(string)`. This is primarily
useful in implementing other function-related traits, but it also
provides a way to get function types to use in is expressions for
testing or providing examples.

In terms of functionality, ToFunctionType is a replacement for
std.traits.FunctionTypeOf. FunctionTypeOf attempts to operate on
anything that's "callable" (both types and symbols), which makes it a
bit of a mess (and which actually cannot work in some cases due to
templated types or functions not having been instantiated). And looking
over where it's used in std.traits, it's usually used on types anyway.

So, ToFunctionType operates exclusively on types. The change in name is
because FunctionTypeOf definitely sounds like a trait that's operating
on a symbol to get its type (even if it also accepts types), whereas
ToFunctionType sounds much more like it's converting the given type,
which is what it's doing.