Fix #10699 - std.format with position with omitted second number does not work
This is part 2 of the fix of #10699.

After the previous commit, the invocation
  format ("%1$x %+2:$d", 11, 22, 33, 44, 55, 66, 77, 88, 99)
produces an exception
  std.format.FormatException@.../std/format/package.d(785): Orphan format arguments: args[65535..9]
apparently because the function formattedWrite returned 65535,
where 9 would be the correct return value. Namely,
  https://dlang.org/library/std/format/write/formatted_write.html
says that the function formattedWrite returns
  "The index of the last argument that was formatted."

The fix does this by correcting the value of `currentArg` at the end of the
`foreach (i, Tunused; Args)` loop. There is no need to update `currentArg`
right before this loop, nor at each iteration of this loop. A single update
of `currentArg`, when the loop is left (through a `goto` or `break` statement),
is sufficient.

Note also that in the assignment `currentArg = i + 1;`, `i + 1` is a constant
expression, as can be seen by compiling the test case with `gdc -S -g`.