Add defaultInit to phobos.sys.traits. (#10842)
This is a new symbol and has no equivalent in std.traits.

The idea here is that in most circumstances, generic code should not use
the init value of a type - either because the init value bypasses
default initialization not working, or because the init value is not
actually the default initialized value for the type (which happens with
non-static nested structs).

Some folks use T() to bypass that problem, because it will not compile
if default initialization doesn't work, and it will properly initialize
the context pointer of non-static nested structs (whereas T.init will
leave the context pointers null).

However, this does not work in generic code, because it's legal to
declare static opCall on structs, and that will hijack T() (and isn't
even guaranteed to return a T, let alone the default-initialized value
of T).

So, defaultInit provides a way to get the default-initialized value of
a type within generic code while not compiling when the type cannot
actually be default-initialized.