phobos#10826 explicitly address at origin in Q1 in beta
This modifies the beta function so that it explicitly addresses the case
where the input parameters (x,y) are (+0,+0), i.e. next to the origin in
the first quadrant. This is already implicitly handled, but explicitly
handling it will be make the algoritm clearer when it is extended to
address the cases where x is very small and y is very small or very
large.

Since the cases are numbered, I had to renumber the cases. (Maybe I
shouldn't have numbered them originally.)