phobos#10826 address Q1 cases where near origin or near infinity (#10890)
* phobos#10826 explicitly address at origin in Q1 in beta

This modifies the beta function so that it explicitly addresses the case
where the input parameters (x,y) are (+0,+0), i.e. next to the origin in
the first quadrant. This is already implicitly handled, but explicitly
handling it will be make the algoritm clearer when it is extended to
address the cases where x is very small and y is very small or very
large.

Since the cases are numbered, I had to renumber the cases. (Maybe I
shouldn't have numbered them originally.)

* phobos#10826 support Q1, near origin or very large

This adds support for the cases when (x,y) is in the first quadrant
and  near the origin or the larger of x and y is very large

* phobos#10826: change NaN handling behavior of beta

This changes the NaN handling behavior of beta to be the same as that of
the + operator, i.e., when both input arguments are NaN, the one with
the larger payload is returned.

* phobos#10826: apply symmetry to entire beta alogrithm