10889: handle reversal truncation issue in betaIncomplete
When x > 1/b and x is larger than the mean, the complement is computed
instead. x -> 1-x. If prior to reversal, x < real.epsilon, 1-x = 1, so
x = 1 after reversal. Since algorthm requires x < 1, the computation of
complement fails. To resolve this failure, when the reversal causes
x -> 1, use x -> nextDown(1-x) instead.

This fix resolved a problematic case for betaIncompleteInv that is
tracked by a unittest. The unittest was updated.