10889: clamp betaDistPowerSeries
The Beta CDF has a hard upper limit of 1. Due to round-off error, the
result of betaDistPowerSeries can be very slightly larger than 1 when
it should be 1. In this case, a value of 1 is returned. To prevent
clamping from hiding a more serious bug, if the result would be greater
than 1 + 2 epsilon, NaN is returned to indicated the computation failed.