10889 extended domain coverage of betaIncomplete and betaIncompleteCompl (#10909)
* 10889: betaIncomplete NaN handling

Bring std.mathspecial.betaIncomplete's handling of NaN parameter values
in line with how D's operators like operator!"+" handle them. I.e., the
NaN with the largest payload is returned.

* 10889: Added support for degenerate cases to betaIncomplete

Support was added to betaIncomplete and betaIncompleteCompl
for the cases when a or b is +0 or infinity.

* 10889 betaIncomplete a=b, x=.5 special case

* 10889: Convert betaIncomplete to use beta

* 10889: handle reversal truncation issue in betaIncomplete

When x > 1/b and x is larger than the mean, the complement is computed
instead. x -> 1-x. If prior to reversal, x < real.epsilon, 1-x = 1, so
x = 1 after reversal. Since algorthm requires x < 1, the computation of
complement fails. To resolve this failure, when the reversal causes
x -> 1, use x -> nextDown(1-x) instead.

This fix resolved a problematic case for betaIncompleteInv that is
tracked by a unittest. The unittest was updated.

* 10889: Extracted logBeta into a function

* 10889: factor constant out of function for later use

* 10889: optimized logBeta

Combine the Stirling approximations of the individual log gamma terms
when one of them is large and the difference between the arguments is
significant. This reduces the impact of subtracting the two larger
terms.

* 10889: betaDistPowerSeries handle a*s approx eq 1

In betaDistPowerSeries, when s*a is close to 1, extreme loss of
precision can occur when combining terms in log space. This fix resolves
this issue.

* 10889: clamp betaDistPowerSeries

The Beta CDF has a hard upper limit of 1. Due to round-off error, the
result of betaDistPowerSeries can be very slightly larger than 1 when
it should be 1. In this case, a value of 1 is returned. To prevent
clamping from hiding a more serious bug, if the result would be greater
than 1 + 2 epsilon, NaN is returned to indicated the computation failed.